+++ nodebackend/src/routes/api.ts
@@ -14,6 +14,7 @@ import et0Router from './et0Route.js';
 import currentWeatherRouter from './currentWeatherRoute.js';
 import nextScheduleRouter from './nextScheduleRoute.js';
 import irrigationRouter from './irrigationRoute.js';
+import cloudsRouter from './cloudsRoute.js';
 
 const router = express.Router();
 
@@ -30,5 +31,6 @@ router.use('/et0', apiLimiter, et0Router);
 router.use('/weather', apiLimiter, currentWeatherRouter);
 router.use('/schedule', apiLimiter, nextScheduleRouter);
 router.use('/irrigation', apiLimiter, irrigationRouter);
+router.use('/clouds', apiLimiter, cloudsRouter);
 
 export default router;